requirements.txt
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-chroma
langchain-groq
langchain
requests
PyPDF2
pypdf
beautifulsoup4
tqdm
sentence-transformers
numpy
faiss-cpu
python-dotenv